+++ src/ngio/iterators/_segmentation.py
@@ -79,6 +79,11 @@ class SegmentationIterator(AbstractIteratorBuilder[np.ndarray, da.Array]):
             "output_transforms": self._output_transforms,
         }
 
+    @property
+    def output_image(self) -> Label:
+        """The label this iterator writes to."""
+        return self._output
+
     def build_numpy_getter(self, roi: Roi) -> DataGetterProtocol[np.ndarray]:
         return NumpyRoiGetter(
             zarr_array=self._input.zarr_array,